Move `build()` function to debug_cli.py
This function backs the `build`-command which is only available in debug
anyways.

Also moves `build_bucket()` to utils.py.
